@@ -4,11 +4,11 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { AuthModule } from './modules/auth/auth.module';
-import { InterviewModule } from './modules/interview/interview.module';
+import { DashboardModule } from './modules/dashboard/dashboard.module';
 
 @NgModule({
   declarations: [AppComponent],
-  imports: [BrowserModule, AppRoutingModule, AuthModule, InterviewModule],
+  imports: [BrowserModule, AppRoutingModule, AuthModule, DashboardModule],
   providers: [],
   bootstrap: [AppComponent],
 })
